+++ DocuSignSDK.xcframework/ios-arm64_i386_x86_64-simulator/DocuSignSDK.framework/Headers/DSESTelemetryEventsResponse.h
@@ -1,9 +1,0 @@
-#import <DocuSignSDK/DSESObject.h>
-#import <DocuSignSDK/DSESFailedTelemetryEvent.h>
-
-@interface DSESTelemetryEventsResponse : DSESObject
-
-@property (nonatomic, copy) NSArray<DSESFailedTelemetryEvent> *failedEvents;
-
-@end
-